Refactor: UserNumberMaker 클래스 별도의 .java 파일로 분리
UserNumberMaker 클래스를 별도의 .java 파일 형태로 분리
